add basic pie chart functionality to profile page
