+++ app/build.gradle
@@ -55,9 +55,7 @@ dependencies {
     implementation 'com.android.support:palette-v7:28.0.+'
     // CardView
     implementation 'com.android.support:cardview-v7:28.0.0'
-    // MultiAdapter
-    implementation 'dev.ahamed.mva2:adapter:2.0.0-alpha01'
-    implementation 'dev.ahamed.mva2:ext-databinding:2.0.0-alpha01'   // DataBinding
-    implementation 'dev.ahamed.mva2:ext-decorator:2.0.0-alpha01'     // Decorators
-    implementation 'dev.ahamed.mva2:ext-diffutil-rx:2.0.0-alpha01'   // RxDiffUtil
+    // Piechart
+    implementation 'com.github.PhilJay:MPAndroidChart:v3.1.0'
+
 }
